Create dedicated RootPitchOption component
Another step towards making the menu's as composable as possible.
