+++ HarpGuru/OverlayMenu/CovariantMenu/components/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import {HarpStrata} from 'harpstrata'
 
 import type {CovariantMenuProps} from '../types'
-import {nudgeHarpStrataByRootPitch} from '../nudgeHarpStrataByRootPitch'
+import { RootPitchOption } from '../../RootPitchOption'
 import { PozitionOption } from '../../PozitionOption'
 import { OptionContainer } from '../../OptionContainer'
 import {MenuContainer} from '../../MenuContainer'
@@ -10,15 +10,6 @@ import {nudgeDisplayMode} from '../../LayoutMenu/nudgeDisplayMode'
 import { HarpKeyOption } from '../../HarpKeyOption'
 import {DisplayModes} from '../../../HarpFace'
 
-
-type FullNudgeFunction = (arg0: HarpStrata, arg1: 'UP' | 'DOWN', arg2: DisplayModes) => HarpStrata
-type PartiallyAppliedNudgeFunction = (arg0: 'UP' | 'DOWN') => HarpStrata
-
-const partiallyApplyCovariantNudgeFunction = (nudgeFunction: FullNudgeFunction, activeHarpStrata: HarpStrata, activeDisplayMode: DisplayModes): PartiallyAppliedNudgeFunction => {
-  return (direction: 'UP' | 'DOWN') => {
-    return nudgeFunction(activeHarpStrata, direction, activeDisplayMode)
-  }
-}
 
 const getPartiallyAppliedDisplayModeNudgeFunction = (activeHarpStrata: HarpStrata, displayMode: DisplayModes, setActiveDisplayMode: (arg0: DisplayModes) => void) => {
   return (): HarpStrata => {
@@ -28,14 +19,6 @@ const getPartiallyAppliedDisplayModeNudgeFunction = (activeHarpStrata: HarpStrat
 
 export const CovariantMenu = (props: CovariantMenuProps): React.ReactElement => {
   const { activeHarpStrata, setActiveHarpStrata, activeDisplayMode, setActiveDisplayMode } = props
-  const { rootPitchId } = activeHarpStrata
-
-  const rootPitchOptionContainerProps = {
-    title: 'Position Key',
-    optionId: rootPitchId,
-    nudgeFunction: partiallyApplyCovariantNudgeFunction(nudgeHarpStrataByRootPitch, activeHarpStrata, activeDisplayMode),
-    setActiveHarpStrata,
-  }
 
   const displayModeOptionContainerProps = {
     title: 'Display',
@@ -48,7 +31,7 @@ export const CovariantMenu = (props: CovariantMenuProps): React.ReactElement => 
     <MenuContainer>
       <HarpKeyOption {...props} />
       <PozitionOption {...props} />
-      <OptionContainer {...rootPitchOptionContainerProps}/>
+      <RootPitchOption {...props} />
       <OptionContainer {...displayModeOptionContainerProps}/>
     </MenuContainer>
   )
